Allow syntax for specifying alignment
This is the first step in allowing more flexible horizontal alignment
specification in the visual syntax. The idea is that "<" will imply left
alignment, ">" right alignment, and "<>" center alignment. This syntax
can be used after the width of a view. Note that alignment doesn't mean
anything if a flexible width is given.

Examples:

"[view(90)<]" - left align
"[view(90)>]" - right align
"[view(90)<>]" - center align

Also note that this commit just allows for the syntax to be used without
crashing, but it does not actually take it into account when creating
constraints.