Implement parsing of horizontal left-right alignment
This commit makes it so a view's horizontal alignment can be specified
using "<" and ">" symbols. These symbols indicate whether or not to pin
the view to the view on its left or right. For example:

[view1(30)<][view2(40)<][view3(50)>]

Indicates to pin view1 to the superview's leading edge, pin view2 to
view1, and pin view3 to the superview's trailing edge. Not that similar
to how the user must be conscious of correctly specifying fixed and
dynamic widths and heights, they must also be conscious of specifying
enough horizontal constraints. For example:

[view1(30)<][view2(40)][view3(50)>]

view2 here is not specified to be pinned to view1 or view3, so its
placement is undertermined.

Also note that a layout such as:

[view1(30)<][view2][view3(50)>]

will also have undertmined behavior even though view2 has dynamic width.
This because the correct syntax for this layout would be:

[view1(30)][view2][view3(50)]

The "<" and ">" specifiers are not needed.