Fix bug where too many constraints added
There was a bug where if a view was being created with one row and there
was a row height specified, a constraint to the bottom of the container
view was still added even though it isn't needed. This caused a problem
when the user would expect the row to remain the same height even if the
container view height was expanded.

This bug was fixed by not adding the bottom constraint if the row known
to have a fixed height.