Fix bug where dynamically sized cells were too small
There was a bug where if an element had a fixed height we did not add a
bottom constraint at all, which could cause dynamically sized cells to
have a height smaller than one if its subviews with a fixed height. This
was done because we wanted to call fixed height views to be in views
that had unfixed heights.

To accommodate both cases, we not add a bottom constraint with a >=
relation for items with a fixed height instead of no constraint at all.