Fix bug where dynamic view on top of fixed view didn't resize
This commit fixes a bug where if a dynamically sized view was placed on
top of a fixed view, it would not grow in size as expected. This bug was
occurring because of a >= constraint being improperly added to the
bottom of the fixed view. We only want to add the >= relation if the
view above the bottom view is of fixed height or the bottom view is the
only view (there is only one row). This is because a user would expect
fixing all view heights would keep the views at the top of the super
view.