Allow custom horizontal spacing when centering items
This commit makes it so when items are centered, their horizontal
spacing is preserved if they were set to values other than the default.

This change also uncovered a bug where custom horizontal spacing
wouldn't work if centering or pinning syntax was used ("<", ">", "<>").
This was because the regex that matched spacing patterns forgot to
include angled brackets in it as a non capturing group.