Add convenience macro for creating variable bindings
This commit adds a macro that wraps NSDictionaryOfVariableBindings() to
allow the passing of `self.<variable_name>` arguments to be converted
from having "self.<varaible_name>" key names to just having
"<variable_name>" keys. For example, passing in `self.someView` will
result in a mapping of "someView" to `self.someView` instead of
"self.someView" to `self.someView`.