Fix bug for specifying row trailing spacing with fixed height
There was a bug where if a custom trailing spacing was specified and a
fixed height was specified the custom trailing spacing wouldn't get
applied. This affected visual formats such as:

`[view]-|(20)`

In this case, the height constraint would get applied but not the
trailing spacing to the superview's right edge. The bug occurred because
the regex for parsing out item spacing didn't account for a trailing
float inside of a parenthesis. The fix for this was to remove the height
constraint string (if there is one) before parsing the string for item
spacing.

The pattern used for parsing out the row height was also simplified.